Display web cam stream
Upon page load, and with the user's permission, the user's web cam will
be turned on and the player will stream the video to the canvas.

Hid the player (top right webcam window).

Added simple CSS to the photobooth class which houses the canvas to have
the width be that of the window.