Video feature - Interact: getting pixel colors
This is the beginning of the ability for the user to wave at certain
areas (using the web cam) and interact in such fashion.
Currently, the top left corner of the web cam has an invisible square.
A function gets the average pixel colors (rbg).
Goal: if the average is that of a pink posted note color when waving it
in that corner, the do -whatever-.