Video feature - Interact: pixel interaction success
Pixel in top right corner are monitored and function utters a message
when pixels are in the range of pink colors.
The colors, message, and selection of pixels are chosen for testing and
are subject to change.
At this point they are successful; a user can flash a pink posted note
and computer speakers will utter 'i saw that'.