Video feature - Calibration
User can now calibrate the top right corner of the webcam.
By placing his/her finger in the corner and (for now) click on calibrate
the user will now have set his average finger pixel colors which when
waved at that corner can trigger something (TBD).
The calibration's range of pixels has been set for testing and is
subject to change.
PS: This is awesome!!